did tasks

+++ 14-binary_tree_balance.c
@@ -12,5 +12,5 @@ int binary_tree_balance(const binary_tree_t *tree)
   if (tree == NULL)
     return (0);
 
-  return (binary_tree_leaves(tree->left) - binary_tree_leaves(tree->right));
+  return (int) (binary_tree_leaves(tree->left) - binary_tree_leaves(tree->right));
 }
